feat(Auditor): code cleanups and minor fixes
